+++ ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: kmuhlbau <kmuhlbau@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/15 23:17:19 by kmuhlbau          #+#    #+#             */
-/*   Updated: 2024/09/15 23:27:24 by kmuhlbau         ###   ########.fr       */
+/*   Updated: 2024/10/07 23:10:58 by kmuhlbau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *b, int c, int len)
+#include "libft.h"
+
+void	*ft_memset(void *b, int c, size_t len)
 {
-	int				i;
+	unsigned int	i;
 	unsigned char	*ptr;
 
 	ptr = b;
